+++ spec/llm_cost_tracker/generator_availability_spec.rb
@@ -16,6 +16,7 @@ RSpec.describe "llm_cost_tracker generators" do
       "llm_cost_tracker:install" => "LlmCostTracker::Generators::InstallGenerator",
       "llm_cost_tracker:add_ingestion" => "LlmCostTracker::Generators::AddIngestionGenerator",
       "llm_cost_tracker:add_call_rollups" => "LlmCostTracker::Generators::AddCallRollupsGenerator",
+      "llm_cost_tracker:add_capture_dimensions" => "LlmCostTracker::Generators::AddCaptureDimensionsGenerator",
       "llm_cost_tracker:add_latency_ms" => "LlmCostTracker::Generators::AddLatencyMsGenerator",
       "llm_cost_tracker:add_streaming" => "LlmCostTracker::Generators::AddStreamingGenerator",
       "llm_cost_tracker:add_provider_response_id" => "LlmCostTracker::Generators::AddProviderResponseIdGenerator",